fixed coding tester config. fixed compile warnings.
Former-commit-id: 9278156363925634a260786fe5096ec8711e943c

+++ trunk/src/communicator/communicator.cc
@@ -243,7 +243,8 @@ void Communicator::waitForMessage() {
 						threadPools[msgType].schedule(
 								boost::bind(&Communicator::dispatch, this, buf,
 										p->first, 0));
-						debug("Add Thread Pool [%s] %d/%d/%d\n",EnumToString::toString(msgType),threadPools[msgType].active(),threadPools[msgType].pending(),threadPools[msgType].size());
+						debug("Add Thread Pool [%s] %d/%d/%d\n",
+								EnumToString::toString(msgType), (int)threadPools[msgType].active(), (int)threadPools[msgType].pending(), (int)threadPools[msgType].size());
 
 #else
 						dispatch(buf, p->first);
@@ -285,20 +286,20 @@ void Communicator::addMessage(Message* message, bool expectReply) {
 	}
 
 	// add message to outMessageQueue
-	switch(message->getMsgHeader().protocolMsgType) {
-		case OBJECT_DATA:
-		case SEGMENT_DATA:
+	switch (message->getMsgHeader().protocolMsgType) {
+	case OBJECT_DATA:
+	case SEGMENT_DATA:
 #ifdef USE_LOWLOCK_QUEUE
-			_outDataQueue.push(message);
+		_outDataQueue.push(message);
 #else
-			_outDataQueue.push(message);
+		_outDataQueue.push(message);
 #endif
-			break;
-		default:
+		break;
+	default:
 #ifdef USE_LOWLOCK_QUEUE
-			_outMessageQueue.push(message); // must be at the end of function
+		_outMessageQueue.push(message); // must be at the end of function
 #else
-			_outMessageQueue.push(message); // must be at the end of function
+				_outMessageQueue.push(message); // must be at the end of function
 #endif
 	}
 
@@ -421,7 +422,7 @@ void Communicator::disconnectAndRemove(uint32_t sockfd) {
 		debug("Connection erased for sockfd = %" PRIu32 "\n", sockfd);
 	} else {
 		cerr << "Connection not found, cannot remove connection" << endl;
-		exit (-1);
+		exit(-1);
 	}
 
 }
@@ -695,8 +696,8 @@ void Communicator::connectToComponents(vector<Component> componentList) {
 	}
 }
 
-void Communicator::connectToMyself(string ip, uint16_t port, ComponentType
-	type) {
+void Communicator::connectToMyself(string ip, uint16_t port,
+		ComponentType type) {
 	uint32_t sockfd = connectAndAdd(ip, port, type);
 	requestHandshake(sockfd, _componentId, _componentType);
 }
